requirements.txt
cassandra-driver
pep8
pandas
jupyter
